Split AudioPlayer in subfiles
